requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
bootstrap-admin==0.4.4
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.2
django-admin-volt==1.0.3
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-grappelli==3.0.3
django-js-asset==2.0.0
django-nested-admin==4.0.0
django-on-heroku==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
promise==2.3
psycopg2-binary==2.9.5
python-dateutil==1.5
python-monkey-business==1.0.0
pytz==2022.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
urllib3==1.26.12
whitenoise==6.2.0
